mycli/template/requirements.txt
python-dotenv
mysql-connector-python
pydantic